+++ R/plot.dispRity_fun.R
@@ -421,7 +421,7 @@ plot.randtest <- function (data_sub, nclass = 10, coeff = 1, ...) {
     legend("topleft", bty = "n", legend = c("p-value", round(data_sub$pvalue, 5)), cex = 0.7, adj = 0.2)
 }
 
-# Plotting model tests results
+## Plotting model tests results
 plot.model.test.support <- function(data, col, ylab, ylim, ...) {
 
     ## Extracting the weighted aicc
@@ -433,6 +433,88 @@ plot.model.test.support <- function(data, col, ylab, ylim, ...) {
     ## Plot
     plotcoords <- graphics::barplot(ordered_aic, col = col, ylim = ylim, ylab = ylab, ...)
 }
+
+## Plotting a space preview
+plot.preview <- function(data, dimensions, xlab, ylab, ylim, col, ...) {
+
+    ## The "ggplot" colours
+    gg.color.hue <- function(n) {
+        grDevices::hcl(h = seq(15, 375, length = n + 1), l = 65, c = 100)[1:n]
+    }
+
+    ## Capturing the dots options
+    plot_args <- list(...)
+
+    ## Setting the dimensions
+    plot_args$x <- data$matrix[, dimensions[1]]
+    plot_args$y <- data$matrix[, dimensions[2]]
+
+    ## Getting the loadings
+    loading <- apply(data$matrix, 2, var)
+    loading <- round(loading/sum(loading)*100, 2)
+
+    ## Setting the labels
+    if(missing(xlab)) {
+        plot_args$xlab <- paste0("Dimension ", dimensions[1], " (", loading[dimensions[1]], "%)")
+    } else {
+        plot_args$xlab <- xlab
+    }
+    if(missing(ylab)) {
+        plot_args$ylab <- paste0("Dimension ", dimensions[2], " (", loading[dimensions[2]], "%)")
+    } else {
+        plot_args$ylab <- ylab
+    }
+
+    ## Setting plot limits
+    plot_lim <- range(as.vector(c(data$matrix[, dimensions])))
+    if(is.null(plot_args$xlim)) {
+        plot_args$xlim <- plot_lim
+    }
+    if(missing(ylim)) {
+        plot_args$ylim <- plot_lim
+    } else {
+        plot_args$ylim <- ylim
+    }
+
+    ## Setting the pch
+    if(is.null(plot_args$pch)) {
+        plot_args$pch <- 19
+    }
+
+    ## Get the number of colour groups
+    n_groups <- length(data$subsets)
+
+    ## Setting the colours
+    if(missing(col)) {
+        if(n_groups == 1) {
+            plot_args$col <- "black"
+        } else {
+            plot_args$col <- gg.color.hue(n_groups)
+        }
+    } else {
+        plot_args$col <- col
+    }
+
+    ## Make a colour classifier
+    if(n_groups > 1) {
+        classifier <- rep(NA, nrow(data$matrix))
+        for(class in 1:n_groups) {
+            classifier[data$subsets[[class]]$elements[,1]] <- class
+        }
+        plot_args$col <- plot_args$col[classifier]
+    }
+
+    ## Plot the results
+    do.call(plot, plot_args)
+    if(n_groups > 1) {
+        legend("topright", legend = names(data$subsets), col = unique(plot_args$col), pch = plot_args$pch, cex = 0.666)
+    }
+
+    ## Return invisible
+    return(invisible())
+}
+
+
 
 # ~~~~~~~~~~
 # sequential.test plots